webhook_listener/requirements.txt
aiohttp
azure-functions
azure-identity
azure-keyvault-secrets
azure-servicebus